+++ bidsme/Modules/EEG/EEG.py
@@ -302,15 +302,18 @@ class EEG(baseModule):
         if self.TableChannels is not None:
             self.TableChannels.to_csv(dest_base + "_channels.tsv",
                                       sep="\t", na_rep="n/a",
-                                      header=True, index=True)
+                                      header=True, index=True,
+                                      line_terminator="\n")
         if self.TableEvents is not None:
             self.TableEvents.to_csv(dest_base + "_events.tsv",
                                     sep="\t", na_rep="n/a",
-                                    header=True, index=True)
+                                    header=True, index=True,
+                                    line_terminator="\n")
         if self.TableElectrodes is not None:
             self.TableElectrodes.to_csv(dest_base + "_electrodes.tsv",
                                         sep="\t", na_rep="n/a",
-                                        header=True, index=True)
+                                        header=True, index=True,
+                                        line_terminator="\n")
         shutil.copy2(self.currentFile(), destination)
 
     def _copy_bidsified(self, directory: str,
@@ -351,7 +354,8 @@ class EEG(baseModule):
             self.TableChannels.to_csv(dest_base + "_channels.tsv",
                                       columns=active,
                                       sep="\t", na_rep="n/a",
-                                      header=True, index=True)
+                                      header=True, index=True,
+                                      line_terminator="\n")
             self._chan_BIDS.DumpDefinitions(dest_base + "_channels.json")
 
         if self.TableEvents is not None and\
@@ -369,7 +373,8 @@ class EEG(baseModule):
             self.TableEvents.to_csv(dest_base + "_events.tsv",
                                     columns=active,
                                     sep="\t", na_rep="n/a",
-                                    header=True, index=True)
+                                    header=True, index=True,
+                                    line_terminator="\n")
             self._task_BIDS.DumpDefinitions(dest_base + "_events.json")
 
         if self.TableElectrodes is not None and\
@@ -387,7 +392,8 @@ class EEG(baseModule):
             self.TableElectrodes.to_csv(dest_base + "_events.tsv",
                                         columns=active,
                                         sep="\t", na_rep="n/a",
-                                        header=True, index=True)
+                                        header=True, index=True,
+                                        line_terminator="\n")
             self._elec_BIDS.DumpDefinitions(dest_base + "_events.json")
 
     @abstractmethod