+++ bidsme/bidsMeta/__init__.py
@@ -1,4 +1,5 @@
 from .BidsSession import BidsSession
 from .BidsMeta import MetaField, BIDSfieldLibrary
+from .BidsTable import BidsTable
 
-__all__ = ["BidsSession", "MetaField", "BIDSfieldLibrary"]
+__all__ = ["BidsSession", "MetaField", "BIDSfieldLibrary", "BidsTable"]