+++ bidsme/tools/paths.py
@@ -42,9 +42,10 @@ app = os.path.splitext(os.path.basename(sys.argv[0]))[0]
 
 local = os.getcwd()
 installation = os.path.normpath(
-        os.path.join(os.path.dirname(__file__), "../.."))
+        os.path.join(os.path.dirname(__file__), "..", ".."))
 
 heuristics = os.path.join(installation, "heuristics")
+templates = os.path.join(installation, "bidsme", "table_templates")
 
 config = appdirs.user_data_dir(app, user)