+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+### Changed
+  - participants.tsv table is managed using bidsMeta.BidsTable class
+
+
 ## [1.3.5r2] - 2021-03-10
 
 ### Fixed